+++ packages/ibm-products/src/components/Datagrid/Datagrid.stories.js
@@ -7,7 +7,7 @@
  */
 
 import React, { useState, useEffect } from 'react';
-import { range, makeData, newPersonWithTwoLines } from './utils/makeData';
+import { makeData } from './utils/makeData';
 import { getStoryTitle } from '../../global/js/utils/story-helper';
 import { action } from '@storybook/addon-actions';
 import { Activity, Add } from '@carbon/react/icons';
@@ -17,7 +17,6 @@ import {
   Datagrid,
   useDatagrid,
   useInfiniteScroll,
-  useRowIsMouseOver,
   useSelectRows,
   useSortableColumns,
   useDisableSelectRows,
@@ -33,7 +32,8 @@ import styles from './_storybook-styles.scss';
 import { DatagridActions } from './utils/DatagridActions';
 import { DatagridPagination } from './utils/DatagridPagination';
 import { Wrapper } from './utils/Wrapper';
-import { pkg } from '../../settings';
+import * as HeaderStory from './Datagrid.stories/Header/Header.stories';
+import * as ColumnAlignmentStory from './Datagrid.stories/ColumnAlignment/ColumnAlignment.stories';
 import { DocsPage } from './Datagrid.docs-page';
 
 export default {
@@ -54,6 +54,7 @@ export default {
       },
     },
   },
+  excludeStories: ['getBatchActions'],
 };
 
 const getColumns = (rows) => {
@@ -234,10 +235,6 @@ export const InfiniteScroll = () => {
     fetchData();
   }, []);
 
-  pkg._silenceWarnings(false); // warnings are ordinarily silenced in storybook, add this to test.
-  pkg.feature['Datagrid.useInfiniteScroll'] = true;
-  pkg._silenceWarnings(true);
-
   const datagridState = useDatagrid(
     {
       columns,
@@ -257,11 +254,8 @@ export const InfiniteScroll = () => {
     </Wrapper>
   );
 };
-InfiniteScroll.args = {
-  featureFlags: ['Datagrid.useInfiniteScroll'],
-};
 
-export const TenThousandEntries = () => {
+export const WithVirtualizedData = () => {
   const [data] = useState(makeData(10000));
   const columns = React.useMemo(() => getColumns(data), []);
   const datagridState = useDatagrid(
@@ -275,7 +269,7 @@ export const TenThousandEntries = () => {
   return <Datagrid datagridState={{ ...datagridState }} />;
 };
 
-export const WithPagination = () => {
+export const Pagination = () => {
   const [data] = useState(makeData(100));
   const columns = React.useMemo(() => getColumns(data), []);
   const datagridState = useDatagrid({
@@ -291,36 +285,8 @@ export const WithPagination = () => {
   return <Datagrid datagridState={{ ...datagridState }} />;
 };
 
-export const IsHoverOnRow = () => {
-  const [data] = useState(makeData(10));
-  const Cell = ({ row }) => {
-    if (row.isMouseOver) {
-      return 'yes hovering!';
-    }
-    return '';
-  };
-  const columns = React.useMemo(
-    () => [
-      ...getColumns(data).slice(0, 3),
-      {
-        Header: 'Is hover on row?',
-        id: 'isHoveringColumn',
-        disableSortBy: true,
-        Cell,
-      },
-    ],
-    []
-  );
-  const datagridState = useDatagrid(
-    {
-      columns,
-      data,
-    },
-    useRowIsMouseOver
-  );
-
-  return <Datagrid datagridState={{ ...datagridState }} />;
-};
+export const Header = HeaderStory.HeaderBasicUsageStory;
+export const ColumnAlignment = ColumnAlignmentStory.ColumnAlignmentStory;
 
 export const SelectableRow = () => {
   const [data] = useState(makeData(10));
@@ -450,7 +416,7 @@ const DatagridBatchActions = (datagridState) => {
   );
 };
 
-const getBatchActions = () => {
+export const getBatchActions = () => {
   return [
     {
       label: 'Duplicate',
@@ -484,10 +450,6 @@ const getBatchActions = () => {
 
 export const BatchActions = () => {
   const [data] = useState(makeData(10));
-
-  pkg._silenceWarnings(false); // warnings are ordinarily silenced in storybook, add this to test.
-  pkg.feature['Datagrid.useActionsColumn'] = true;
-  pkg._silenceWarnings(true);
 
   const columns = React.useMemo(
     () => [
@@ -530,9 +492,9 @@ export const BatchActions = () => {
       DatagridBatchActions,
       rowActions: getRowActions(),
       onSelectAllRows: () => console.log('onSelectAll batch action callback'),
+      batchActionMenuButtonLabel: 'More',
     },
     useSelectRows,
-    useSelectAllWithToggle,
     useActionsColumn,
     useStickyColumn
   );
@@ -559,48 +521,8 @@ export const DisableSelectRow = () => {
   return <Datagrid datagridState={{ ...datagridState }} />;
 };
 
-const makeDataWithTwoLines = (length) =>
-  range(length).map(() => newPersonWithTwoLines());
-
-export const TopAlignment = () => {
-  const [data] = useState(makeDataWithTwoLines(10));
-  const columns = React.useMemo(() => getColumns(data).slice(0, 3), []);
-  const datagridState = useDatagrid(
-    {
-      columns,
-      data,
-      verticalAlign: 'top',
-      variableRowHeight: true,
-      rowSize: 'xs',
-      rowSizes: [
-        {
-          value: 'xl',
-        },
-        {
-          value: 'lg',
-        },
-        {
-          value: 'md',
-        },
-        {
-          value: 'xs',
-        },
-      ],
-      DatagridActions,
-      DatagridBatchActions,
-    },
-    useSelectRows
-  );
-
-  return <Datagrid datagridState={{ ...datagridState }} />;
-};
-
 export const FrozenColumns = () => {
   const [data] = useState(makeData(10));
-
-  pkg._silenceWarnings(false); // warnings are ordinarily silenced in storybook, add this to test.
-  pkg.feature['Datagrid.useActionsColumn'] = true;
-  pkg._silenceWarnings(true);
 
   const columns = React.useMemo(
     () => [