+++ packages/ibm-products/src/components/Datagrid/Datagrid/DatagridRow.js
@@ -24,7 +24,15 @@ const rowHeights = {
 
 // eslint-disable-next-line react/prop-types
 const DatagridRow = (datagridState) => {
-  const { row, rowSize, withNestedRows, prepareRow, key } = datagridState;
+  const {
+    row,
+    rowSize,
+    withNestedRows,
+    prepareRow,
+    key,
+    tableId,
+    withExpandedRows,
+  } = datagridState;
 
   const getVisibleNestedRowCount = ({ isExpanded, subRows }) => {
     let size = 0;
@@ -67,7 +75,7 @@ const DatagridRow = (datagridState) => {
 
   const focusRemover = () => {
     const elements = document.querySelectorAll(
-      `.${blockClass}__carbon-row-expanded`
+      `#${tableId} .${blockClass}__carbon-row-expanded`
     );
     elements.forEach((el) => {
       el.classList.remove(`${blockClass}__carbon-row-expanded-hover-active`);
@@ -107,6 +115,15 @@ const DatagridRow = (datagridState) => {
     [`${carbon.prefix}--data-table--selected`]: row.isSelected,
   });
 
+  const setAdditionalRowProps = () => {
+    if (withNestedRows || withExpandedRows) {
+      return {
+        'data-nested-row-id': row.id,
+      };
+    }
+    return {};
+  };
+
   return (
     <React.Fragment key={key}>
       <TableRow
@@ -118,6 +135,7 @@ const DatagridRow = (datagridState) => {
         onFocus={hoverHandler}
         onBlur={focusRemover}
         onKeyUp={handleOnKeyUp}
+        {...setAdditionalRowProps()}
       >
         {row.cells.map((cell, index) => {
           const cellProps = cell.getCellProps();