+++ packages/ibm-products/src/components/Datagrid/Extensions/ColumnCustomization/ColumnCustomization.stories.js
@@ -259,10 +259,6 @@ export const ColumnCustomizationUsageStory = prepareStory(
 const ColumnCustomizationWithFixedColumn = ({ ...args }) => {
   const stickyHeaders = defaultHeader.slice(1, 15);
 
-  pkg._silenceWarnings(false);
-  pkg.feature['Datagrid.useActionsColumn'] = true;
-  pkg._silenceWarnings(true);
-
   const columns = React.useMemo(
     () => [
       {