+++ packages/ibm-products/src/components/Datagrid/Extensions/NestedRows/NestedRows.stories.js
@@ -15,11 +15,9 @@ import {
 } from '../../../../global/js/utils/story-helper';
 import { Datagrid, useDatagrid, useNestedRows } from '../../index';
 import styles from '../../_storybook-styles.scss';
-// import mdx from '../../Datagrid.mdx';
 import { DatagridActions } from '../../utils/DatagridActions';
 import { makeData } from '../../utils/makeData';
 import { ARG_TYPES } from '../../utils/getArgTypes';
-import { pkg } from '../../../../settings';
 import { StoryDocsPage } from '../../../../global/js/utils/StoryDocsPage';
 
 export default {
@@ -183,6 +181,57 @@ const sharedDatagridProps = {
   ],
 };
 
+const nestedRowsControlProps = {
+  gridTitle: sharedDatagridProps.gridTitle,
+  gridDescription: sharedDatagridProps.gridDescription,
+  useDenseHeader: sharedDatagridProps.useDenseHeader,
+  rowSize: sharedDatagridProps.rowSize,
+  rowSizes: sharedDatagridProps.rowSizes,
+  onRowSizeChange: sharedDatagridProps.onRowSizeChange,
+};
+
+const SingleLevelNestedRows = ({ ...args }) => {
+  const columns = React.useMemo(() => defaultHeader, []);
+  const [data] = useState(makeData(10, 2));
+  const datagridState = useDatagrid(
+    {
+      columns,
+      data,
+      DatagridActions,
+      ...args.defaultGridProps,
+    },
+    useNestedRows
+  );
+
+  return <Datagrid datagridState={{ ...datagridState }} />;
+};
+
+const SingleLevelNestedRowsWrapper = ({ ...args }) => {
+  return <SingleLevelNestedRows defaultGridProps={{ ...args }} />;
+};
+
+const singleNestedRowsStoryName = 'With single-level nested rows';
+export const SingleLevelNestedRowsUsageStory = prepareStory(
+  SingleLevelNestedRowsWrapper,
+  {
+    storyName: singleNestedRowsStoryName,
+    argTypes: {
+      gridTitle: ARG_TYPES.gridTitle,
+      gridDescription: ARG_TYPES.gridDescription,
+      useDenseHeader: ARG_TYPES.useDenseHeader,
+      rowSize: ARG_TYPES.rowSize,
+      rowSizes: ARG_TYPES.rowSizes,
+      onRowSizeChange: ARG_TYPES.onRowSizeChange,
+      expanderButtonTitleExpanded: 'Collapse row',
+      expanderButtonTitleCollapsed: 'Expand row',
+    },
+    args: {
+      ...nestedRowsControlProps,
+      featureFlags: ['Datagrid.useNestedRows'],
+    },
+  }
+);
+
 const NestedRows = ({ ...args }) => {
   const columns = React.useMemo(() => defaultHeader, []);
   const [data] = useState(makeData(10, 5, 2, 2));
@@ -196,12 +245,6 @@ const NestedRows = ({ ...args }) => {
     useNestedRows
   );
 
-  // Warnings are ordinarily silenced in storybook, add this to test
-  pkg._silenceWarnings(false);
-  // Enable feature flag for `useNestedRows` hook
-  pkg.feature['Datagrid.useNestedRows'] = true;
-  pkg._silenceWarnings(true);
-
   return <Datagrid datagridState={{ ...datagridState }} />;
 };
 
@@ -209,14 +252,6 @@ const BasicTemplateWrapper = ({ ...args }) => {
   return <NestedRows defaultGridProps={{ ...args }} />;
 };
 
-const nestedRowsControlProps = {
-  gridTitle: sharedDatagridProps.gridTitle,
-  gridDescription: sharedDatagridProps.gridDescription,
-  useDenseHeader: sharedDatagridProps.useDenseHeader,
-  rowSize: sharedDatagridProps.rowSize,
-  rowSizes: sharedDatagridProps.rowSizes,
-  onRowSizeChange: sharedDatagridProps.onRowSizeChange,
-};
 const nestedRowsStoryName = 'With nested rows';
 export const NestedRowsUsageStory = prepareStory(BasicTemplateWrapper, {
   storyName: nestedRowsStoryName,
@@ -232,6 +267,5 @@ export const NestedRowsUsageStory = prepareStory(BasicTemplateWrapper, {
   },
   args: {
     ...nestedRowsControlProps,
-    featureFlags: ['Datagrid.useNestedRows'],
   },
 });