+++ packages/ibm-products/src/components/Datagrid/useFlexResize.js
@@ -1,24 +1,34 @@
-/*
- * Licensed Materials - Property of IBM
- * 5724-Q36
- * (c) Copyright IBM Corp. 2020
- * US Government Users Restricted Rights - Use, duplication or disclosure
- * restricted by GSA ADP Schedule Contract with IBM Corp.
+/**
+ * Copyright IBM Corp. 2020, 2023
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
  */
+
 const useFlexResize = (hooks) => {
-  const spacerColumn = {
+  const spacer = {
     id: 'spacer',
     width: 0,
     disableSortBy: true,
     disableResizing: true,
   };
-  hooks.visibleColumns.push((columns) => [...columns, spacerColumn]);
+  hooks.visibleColumns.push((columns) => {
+    // always move actions and spacer to the end
+    const actionsIdx = columns.findIndex((col) => col.isAction);
+    if (actionsIdx === -1) {
+      return [...columns, spacer];
+    }
+    const cols = [...columns];
+    const actions = cols.splice(actionsIdx, 1)[0];
+    cols.splice(columns.length, 0, spacer, actions);
+    return cols;
+  });
   const changeProps = (props, data) => {
     let { column } = data;
     if (!column && data.cell) {
       column = data.cell.column;
     }
-    if (column.id === spacerColumn.id) {
+    if (column.id === spacer.id) {
       return [props, { style: { flex: '1 1 0' } }];
     }
     return [props, { style: { flex: '0 0 auto' } }];